Added ARM_ALLOW_MULTI_CORE, putting unused cores to sleep
When using a single core environment, the CPU cores 1-3 are
continuously reading a local mailbox register to get a start address.
They are competing for bus time with core 0 what slows it down. And it
breaks the SPI output. At startup the unused cores are putting to sleep.
This was suggested by @rsta2