ltcgenerator.cpp - add ltc!start@hh:mm:ss:ff
the @ operator stops output, set start to the new hh:mm:ss:ff, and then does a start-stop, to timewarp the output to this exact frame.  
After this command, we are stopped, and waiting for a !start.     
This is ideal for scrubbing around on a timeline in an editor.

I thought about using ltc!stop#command, but that would change the stop time, which is not what I am trying to do really, then I thought a new ltc!jump command, but really not worth it.   Then I thought,  swap the '!' command with the '@' command.   What do you think?